File upload refactor
The files that are now uploading are no longer sent as a base64 string and saved into the db. Now, the data is send inside a FormData object.

The files are now saved inside a folder and inside the db is stored path towards the specific folder.

The files are deleted when the product is deleted.